+++ CHANGELOG.md
@@ -17,7 +17,7 @@ The Issue 91 fix brings incompatibility to the past. Class name changed.
 - [Issue 101 : Add I2S and ADC callback handler to the template. ](https://github.com/suikan4github/murasaki/issues/101)
 - [Issue 102 : I2sSearch() should be in separate file ](https://github.com/suikan4github/murasaki/issues/102)
 - [Issue 103 : Restructure the document module hierarchy ](https://github.com/suikan4github/murasaki/issues/103)
-
+- [Issue 107 : Reject STM32H7 from I2sPortAdapter class by assertion ](https://github.com/suikan4github/murasaki/issues/107)
 ### Deprecated
 ### Removed
 ### Fixed